Began work implementing update secrets
