+++ test/test_config.py
@@ -8,17 +8,17 @@ from aws_cdk.aws_ssm import StringParameter
 from moto import mock_secretsmanager, mock_ssm
 
 from cdkutils.config import (
-    AccountIdConfig,
-    AttributeNotFoundException,
-    CommonPipelineConfig,
     ConfigException,
-    GitHubConfig,
-    PersistedAttribute,
+    SsmConfig,
     PipelineConfig,
+    CommonPipelineConfig,
+    AccountIdConfig,
+    GitHubConfig,
     PipIndexConfig,
     ServiceDetails,
-    SsmConfig,
 )
+from cdkutils.persistent_config import PersistedAttribute
+from cdkutils.errors import AttributeNotFoundException
 
 if TYPE_CHECKING:
     from mypy_boto3_secretsmanager.client import SecretsManagerClient
@@ -640,6 +640,32 @@ class PipIndexConfigTest(TestCase):
             secrets_mgr.get_secret_value(SecretId=pip_conf.get_ssm_param_name("username"))
         with self.assertRaises(secrets_mgr.exceptions.ResourceNotFoundException):
             secrets_mgr.get_secret_value(SecretId=pip_conf.get_ssm_param_name("url"))
+
+    @mock_ssm
+    @mock_secretsmanager
+    def test_update_secret(self):
+        """
+        Given a new secret config
+        When the old secret exists on AWS
+        Then the secret is updated
+        """
+        test_username = "foo@metoffice.gov.uk"
+        test_password = "this_is_the_password"
+
+        sm_client = boto3.client("secretsmanager", region_name="eu-west-2")
+        sm_client.create_secret(Name=test_username, SecretString=test_password)
+
+        test_new_password = "new_password"
+
+        pip_conf = PipIndexConfig(self.test_ssm_config, test_username, test_new_password)
+        pip_conf.create_secrets(boto3.Session(region_name="eu-west-2"))
+
+        secrets_mgr: SecretsManagerClient = boto3.client("secretsmanager", region_name="eu-west-2")
+
+        self.assertEqual(
+            pip_conf.password,
+            secrets_mgr.get_secret_value(SecretId=pip_conf.get_secret_name("password"))["SecretString"],
+        )
 
     @mock_ssm
     @mock_secretsmanager
